Introduce cargo-archiver to tar crate dependencies
Designed to be inherited globally along with `archiver`,
`cargo-archiver` will run `cargo vendor` for each package inheriting
`cargo`, and insert the resulting tarball into `$ARCHIVER_OUTDIR` to be
deployed along with the package sources.

Currently does not filter crates by `COPYLEFT_LICENSE_*`, and must be
done in post-processing if desired.